+++ src/server/java/com/softwareverde/bitcoin/server/module/stratum/WorkerShareCallback.java
@@ -1,5 +1,5 @@
 package com.softwareverde.bitcoin.server.module.stratum;
 
 public interface WorkerShareCallback {
-    void onNewWorkerShare(String workerUsername, Integer shareDifficulty);
+    void onNewWorkerShare(String workerUsername, Long shareDifficulty);
 }
